backend/requirements.txt
Django==6.0.4
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.3.0
django-filter==25.2
psycopg2-binary==2.9.9
python-decouple==3.8
drf-spectacular==0.27.0
google-auth==2.25.2
requests==2.32.3
gunicorn==21.2.0
setuptools==82.0.1
